survey_api/requirements.txt
Django==1.10.4
djangorestframework==3.5.3
pymongo==3.4.0

